feat: implement cart functionality with product addition, removal, and quantity editing

+++ database/migrations/2025_06_06_220811_create_produtos_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('produtos', function (Blueprint $table) {
             $table->id(); // codigo de barras
             $table->unsignedBigInteger('loja_id');
-            $table->integer('codigo_barras')->unique();
+            $table->bigInteger('codigo_barras')->unique();
             $table->string('nome')->unique();
             $table->string('marca');
             $table->double('preco_venda');
